pyproject.toml
[project]
name = "abstract-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ray",
    "pandas",
    "pypdf",
    "requests",
    "beautifulsoup4",
    "lxml",
    "langgraph>=1.0.10",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.0.1",
]
